requirements.txt
pytest==8.0.1
selenium==4.17.2
allure-pytest==2.13.5
pre-commit==3.7.1
flake8==7.0.0
